chore: separate index route
